Allow ships heading to be a float (again)

+++ omnirose/curve/forms.py
@@ -5,7 +5,7 @@ from django.forms.widgets import NumberInput
 from .models import Reading
 
 class ReadingForm(forms.Form):
-    ships_head = forms.IntegerField(required=False, widget=NumberInput(attrs={'style': "width: 5em;"}))
-    deviation  = forms.FloatField(required=False,   widget=NumberInput(attrs={'style': "width: 5em;"}))
+    ships_head = forms.FloatField(required=False, widget=NumberInput(attrs={'style': "width: 5em;"}))
+    deviation  = forms.FloatField(required=False, widget=NumberInput(attrs={'style': "width: 5em;"}))
 
 ReadingFormSet = formset_factory(form=ReadingForm)
